uint16_t -> unsigned and uint32_t -> unsigned

+++ src/socket/SocketType.hpp
@@ -19,8 +19,8 @@ enum t_socket_domain { SD_IP4, SD_IP6 };
 enum t_socket_type { ST_TCP, ST_UDP };
 
 typedef int t_fd;
-typedef uint16_t t_port;
-typedef uint32_t t_addressv4;
+typedef unsigned t_port;
+typedef unsigned t_addressv4;
 
 class ISocketLike;
